+++ src/constants.ts
@@ -1,6 +1,7 @@
 // Copyright (c) 2021 Visiosto oy
 // Licensed under the MIT License
 
+export const colorModeKey = 'color-mode';
 export const initialColorModeCSSProp = '--initial-color-mode';
 
 export const formClassErrorMessage = 'error-message';